// src/app/api-client/model/groupsSummaryResponse.ts
/**
 * FastAPI
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GroupSummaryEntry } from './groupSummaryEntry';

/**
 * Enriched overview of all text model groups.
 */
export interface GroupsSummaryResponse {
  groups: Array<GroupSummaryEntry>;
  total_groups: number;
  total_models: number;
  groups_with_families: number;
  groups_with_aliases: number;
  groups_with_issues: number;
}