// src/app/api-client/model/groupHealthIssue.ts
/**
 * FastAPI
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * A single health problem detected for a group.
 */
export interface GroupHealthIssue {
  group_name: string;
  issue_type: string;
  message: string;
  severity?: string;
}